requirements.txt
streamlit>=1.28.0
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.1.0
numpy>=1.26.0
tensorflow>=2.15.0
finnhub-python>=2.4.18
plotly>=5.18.0
matplotlib>=3.8.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
yfinance>=0.2.32
